api/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.10.2
numpy==2.4.2
pandas==3.0.1
PyJWT==2.11.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
